+++ R/as-ggplot.R
@@ -6,6 +6,7 @@
 ##' @param scale scale of the plot to be drawn
 ##' @param hjust horizontal adjustment
 ##' @param vjust vertical adjustment
+##' @param angle angle to rotate plot
 ##' @importFrom ggplot2 ggplot
 ##' @importFrom ggplot2 aes_
 ##' @importFrom ggplot2 geom_blank
@@ -14,16 +15,37 @@
 ##' @importFrom ggplot2 scale_x_continuous
 ##' @importFrom ggplot2 scale_y_continuous
 ##' @return ggplot object
+##' @importFrom grid viewport
 ##' @export
 ##' @examples
 ##' as.ggplot(~barplot(1:10))
 ##' @author Guangchuang Yu
-as.ggplot <- function(plot, scale = 1, hjust = 0, vjust = 0) {
+as.ggplot <- function(plot, scale = 1, hjust = 0, vjust = 0, angle = 0) {
     ## plot_expr <- quo_name(enexpr(plot))
     ## if (is.null(plot)) {
     ##     plot <- as.grob(plot_expr)
     ## }
 
+    if (angle == 0) {
+        return(as.ggplot_internal(plot = plot,
+                                  scale = scale,
+                                  hjust = hjust,
+                                  vjust = vjust))
+    }
+
+    g <- grid2grob(print(as.ggplot_internal(plot),
+                         newpage = TRUE,
+                         vp = viewport(x = .5 + hjust,
+                                       y = .5 + vjust,
+                                       angle = angle,
+                                       width = scale,
+                                       height = scale)
+                         ))
+    as.ggplot(g)
+}
+
+
+as.ggplot_internal <- function(plot, scale = 1, hjust = 0, vjust = 0) {
     ymin <- xmin <- 1 - scale
     xmax <- ymax <- scale