+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggplotify 0.0.5.991
 
-+ `rotate` function to display plot at specific angle (2020-03-16, Mon)
++ angle parameter in `as.ggplot` (2020-03-16, Mon)
 
 # ggplotify 0.0.5